Share one write-through core between the two automation delete paths
